BackendAsistencia/requirements.txt
Django==5.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
reportlab==4.0.7
Pillow==10.1.0
qrcode==7.4.2

